fix kyc data dialog && remove kyc from identity screen
