+++ app/lib/screens/identity_verification_screen.dart
@@ -1,33 +1,21 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:core';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_pkid/flutter_pkid.dart';
 import 'package:http/http.dart';
-import 'package:idenfy_sdk_flutter/idenfy_sdk_flutter.dart';
-import 'package:idenfy_sdk_flutter/models/auto_identification_status.dart';
-import 'package:idenfy_sdk_flutter/models/idenfy_identification_status.dart';
-import 'package:threebotlogin/events/events.dart';
-import 'package:threebotlogin/events/identity_callback_event.dart';
 import 'package:threebotlogin/helpers/globals.dart';
 import 'package:threebotlogin/helpers/kyc_helpers.dart';
 import 'package:threebotlogin/helpers/logger.dart';
 import 'package:threebotlogin/main.dart';
-import 'package:threebotlogin/models/idenfy.dart';
-import 'package:threebotlogin/models/wallet.dart';
 import 'package:threebotlogin/screens/authentication_screen.dart';
-import 'package:threebotlogin/screens/wizard/web_view.dart';
 import 'package:threebotlogin/services/gridproxy_service.dart';
-import 'package:threebotlogin/services/idenfy_service.dart';
 import 'package:threebotlogin/services/identity_service.dart';
 import 'package:threebotlogin/services/open_kyc_service.dart';
 import 'package:threebotlogin/services/pkid_service.dart';
-import 'package:threebotlogin/services/tfchain_service.dart';
 import 'package:threebotlogin/services/tools_service.dart';
 import 'package:threebotlogin/services/shared_preference_service.dart';
-import 'package:threebotlogin/services/wallet_service.dart';
 import 'package:threebotlogin/widgets/custom_dialog.dart';
 import 'package:threebotlogin/widgets/kyc_widget.dart';
 import 'package:threebotlogin/widgets/layout_drawer.dart';
@@ -48,15 +36,11 @@ class _IdentityVerificationScreenState
   String email = '';
   String phone = '';
 
-  String kycLogs = '';
-
   String reference = '';
 
   bool emailVerified = false;
   bool phoneVerified = false;
-  bool identityVerified = false;
 
-  bool isInIdentityProcess = false;
   bool isLoading = false;
 
   bool hidePhoneVerifyButton = false;
@@ -129,14 +113,6 @@ class _IdentityVerificationScreenState
     }
   }
 
-  setIdentityVerified() {
-    if (mounted) {
-      setState(() {
-        identityVerified = Globals().identityVerified.value;
-      });
-    }
-  }
-
   setHidePhoneVerify() {
     if (mounted) {
       setState(() {
@@ -151,7 +127,6 @@ class _IdentityVerificationScreenState
 
     Globals().emailVerified.addListener(setEmailVerified);
     Globals().phoneVerified.addListener(setPhoneVerified);
-    Globals().identityVerified.addListener(setIdentityVerified);
     Globals().hidePhoneButton.addListener(setHidePhoneVerify);
     checkPhoneStatus();
     getUserValues();
@@ -202,17 +177,6 @@ class _IdentityVerificationScreenState
         }
       });
     });
-    getIdentity().then((verificationDate) {
-      setState(() {
-        if (verificationDate['identityName'] != null) {
-          identityVerified = true;
-          setIsIdentityVerified(true);
-        } else {
-          identityVerified = false;
-          setIsIdentityVerified(false);
-        }
-      });
-    });
     getSpending();
   }
 
@@ -239,7 +203,7 @@ class _IdentityVerificationScreenState
         builder: (ctx, snapshot) {
           if (snapshot.connectionState == ConnectionState.done) {
             if (isLoading) {
-              return _pleaseWait();
+              return pleaseWait(context);
             }
 
             return Column(
@@ -253,7 +217,6 @@ class _IdentityVerificationScreenState
                           animation: Listenable.merge([
                             Globals().emailVerified,
                             Globals().phoneVerified,
-                            Globals().identityVerified
                           ]),
                           builder: (BuildContext context, _) {
                             return Column(
@@ -290,7 +253,7 @@ class _IdentityVerificationScreenState
                                 // Step one: verify email
                                 _fillCard(
                                     getCorrectState(1, emailVerified,
-                                        phoneVerified, identityVerified),
+                                        phoneVerified),
                                     1,
                                     email,
                                     Icons.email),
@@ -302,64 +265,19 @@ class _IdentityVerificationScreenState
                                             spending > Globals().spendingLimit))
                                     ? _fillCard(
                                         getCorrectState(2, emailVerified,
-                                            phoneVerified, identityVerified),
+                                            phoneVerified),
                                         2,
                                         phone,
                                         Icons.phone)
                                     : Container(),
                                 customDivider(context: context),
+                                const ListTile(
+                                  leading: Icon(Icons.info),
+                                  title: Text(
+                                    'KYC Verification has been moved to wallet page.'
+                                  ),
+                                ),
 
-                                // Step three: verify identity
-                                (Globals().isOpenKYCEnabled ||
-                                        (Globals().spendingLimit > 0 &&
-                                            spending > Globals().spendingLimit))
-                                    ? _fillCard(
-                                        getCorrectState(3, emailVerified,
-                                            phoneVerified, identityVerified),
-                                        3,
-                                        extract3Bot(doubleName),
-                                        Icons.perm_identity)
-                                    : Container(),
-
-                                Globals().redoIdentityVerification &&
-                                        identityVerified == true
-                                    ? ElevatedButton(
-                                        onPressed: () async {
-                                          // await verifyIdentityProcess(
-                                          //     context: context,
-                                          //     setIdentityProcess: (value) =>
-                                          //         setState(() =>
-                                          //             isInIdentityProcess =
-                                          //                 value),
-                                          //     setLoading: (value) => setState(
-                                          //         () => isLoading = value));
-                                        },
-                                        child: const Text(
-                                            'Redo identity verification'))
-                                    : Container(),
-                                Globals().debugMode == true
-                                    ? ElevatedButton(
-                                        onPressed: () async {
-                                          bool? isEmailVerified =
-                                              await getIsEmailVerified();
-                                          bool? isPhoneVerified =
-                                              await getIsPhoneVerified();
-                                          bool? isIdentityVerified =
-                                              await getIsIdentityVerified();
-
-                                          kycLogs = '';
-                                          kycLogs +=
-                                              'Email verified: $isEmailVerified\n';
-                                          kycLogs +=
-                                              'Phone verified: $isPhoneVerified\n';
-                                          kycLogs +=
-                                              'Identity verified: $isIdentityVerified\n';
-
-                                          setState(() {});
-                                        },
-                                        child: const Text('KYC Status'))
-                                    : Container(),
-                                Text(kycLogs),
                               ],
                             );
                           })
@@ -369,215 +287,8 @@ class _IdentityVerificationScreenState
               ],
             );
           }
-          return _pleaseWait();
+          return pleaseWait(context);
         },
-      ),
-    );
-  }
-
-  termsAndConditionsDialog() {
-    bool isAccepted = false;
-
-    showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext customContext) {
-        return StatefulBuilder(
-          builder: (BuildContext context, StateSetter setState) {
-            return CustomDialog(
-              title: 'Terms and Conditions',
-              type: DialogType.Info,
-              image: Icons.info,
-              widgetDescription: Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    SizedBox(height: MediaQuery.of(context).size.height * 0.01),
-                    RichText(
-                      text: TextSpan(
-                        text:
-                            "As part of the verification process, we utilize iDenfy to verify your identity. Please ensure you review iDenfy's ",
-                        style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                              color: Theme.of(context).colorScheme.onSurface,
-                            ),
-                        children: [
-                          TextSpan(
-                            text: 'Security and Compliance',
-                            style: Theme.of(context)
-                                .textTheme
-                                .bodyMedium!
-                                .copyWith(
-                                  fontWeight: FontWeight.bold,
-                                  color:
-                                      Theme.of(context).colorScheme.onSurface,
-                                ),
-                          ),
-                          TextSpan(
-                            text: ', which include their ',
-                            style: Theme.of(context)
-                                .textTheme
-                                .bodyMedium!
-                                .copyWith(
-                                  color:
-                                      Theme.of(context).colorScheme.onSurface,
-                                ),
-                          ),
-                          TextSpan(
-                            text: 'Terms & Conditions, Privacy Policy,',
-                            style: Theme.of(context)
-                                .textTheme
-                                .bodyMedium!
-                                .copyWith(
-                                  fontWeight: FontWeight.bold,
-                                  color:
-                                      Theme.of(context).colorScheme.onSurface,
-                                ),
-                          ),
-                          TextSpan(
-                            text: ' and other relevant documents.',
-                            style: Theme.of(context)
-                                .textTheme
-                                .bodyMedium!
-                                .copyWith(
-                                  color:
-                                      Theme.of(context).colorScheme.onSurface,
-                                ),
-                          )
-                          //
-                        ],
-                      ),
-                    ),
-                    Row(
-                      children: [
-                        Checkbox(
-                          value: isAccepted,
-                          onChanged: (bool? value) {
-                            setState(() {
-                              isAccepted = value ?? false;
-                            });
-                          },
-                        ),
-                        Expanded(
-                          child: RichText(
-                            text: TextSpan(
-                              text: 'I have read and agreed to ',
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .bodyMedium!
-                                  .copyWith(
-                                    color:
-                                        Theme.of(context).colorScheme.onSurface,
-                                  ),
-                              children: [
-                                TextSpan(
-                                  text: 'iDenfy Terms and Conditions.',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .primary,
-                                        decoration: TextDecoration.underline,
-                                      ),
-                                  recognizer: TapGestureRecognizer()
-                                    ..onTap = () {
-                                      Navigator.push(
-                                        context,
-                                        MaterialPageRoute(
-                                          builder: (context) => const WebView(
-                                            url:
-                                                'https://www.idenfy.com/security/',
-                                            title:
-                                                'iDenfy Terms and Conditions',
-                                          ),
-                                        ),
-                                      );
-                                    },
-                                ),
-                                TextSpan(
-                                  text: '.',
-                                  style: Theme.of(context).textTheme.bodyMedium,
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                      ],
-                    )
-                  ],
-                ),
-              ),
-              actions: [
-                TextButton(
-                  onPressed: () {
-                    Navigator.pop(customContext);
-                  },
-                  child: const Text('Cancel'),
-                ),
-                TextButton(
-                  onPressed: isAccepted
-                      ? () async {
-                          Navigator.pop(customContext);
-                          // await verifyIdentityProcess(
-                          //     context: context,
-                          //     setIdentityProcess: (value) =>
-                          //         setState(() => isInIdentityProcess = value),
-                          //     setLoading: (value) =>
-                          //         setState(() => isLoading = value));
-                        }
-                      : null,
-                  child: Text(
-                    'Continue',
-                    style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                          color: isAccepted
-                              ? Theme.of(context).colorScheme.primary
-                              : Theme.of(context).disabledColor,
-                        ),
-                  ),
-                ),
-              ],
-            );
-          },
-        );
-      },
-    );
-  }
-
-  showAreYouSureToExitDialog() {
-    return showDialog(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext customContext) => CustomDialog(
-        type: DialogType.Warning,
-        image: Icons.warning,
-        title: 'Are you sure',
-        description: 'Are you sure you want to exit the verification process',
-        actions: <Widget>[
-          TextButton(
-            child: const Text('No'),
-            onPressed: () async {
-              Navigator.pop(customContext);
-            },
-          ),
-          TextButton(
-            child: Text(
-              'Yes',
-              style: Theme.of(context)
-                  .textTheme
-                  .bodyMedium!
-                  .copyWith(color: Theme.of(context).colorScheme.warning),
-            ),
-            onPressed: () async {
-              Navigator.pop(customContext);
-              setState(() {
-                isInIdentityProcess = false;
-              });
-            },
-          ),
-        ],
       ),
     );
   }
@@ -626,189 +337,6 @@ class _IdentityVerificationScreenState
         ),
       );
     }
-  }
-
-  Future<void> initIdenfySdk(String token) async {
-    IdenfyIdentificationResult? idenfySDKresult;
-    try {
-      idenfySDKresult = await IdenfySdkFlutter.start(token);
-    } catch (e) {
-      logger.e(e);
-      if (context.mounted) {
-        showDialog(
-          context: context,
-          barrierDismissible: false,
-          builder: (BuildContext dialogContext) => CustomDialog(
-            type: DialogType.Error,
-            image: Icons.close,
-            title: 'Error',
-            description:
-                'Something went wrong. Please contact support if this issue persists.',
-            actions: [
-              TextButton(
-                  onPressed: () {
-                    Navigator.pop(dialogContext);
-                  },
-                  child: const Text('Close'))
-            ],
-          ),
-        );
-      }
-    }
-    await Future.delayed(const Duration(seconds: 5));
-    if (idenfySDKresult != null &&
-        idenfySDKresult.autoIdentificationStatus !=
-            AutoIdentificationStatus.UNVERIFIED) {
-      // await handleIdenfyResponse();
-    }
-  }
-
-  // Future<void> handleIdenfyResponse() async {
-  //   VerificationStatus verificationStatus;
-  //   try {
-  //     final address = await getMyAddress();
-  //     final idenfyServiceUrl = Globals().idenfyServiceUrl;
-  //     verificationStatus = await getVerificationStatus(address: address, idenfyServiceUrl: idenfyServiceUrl);
-  //   } catch (e) {
-  //     setState(() {
-  //       isLoading = false;
-  //     });
-  //     logger.e(e);
-  //     return showDialog(
-  //       context: context,
-  //       builder: (BuildContext context) => CustomDialog(
-  //         type: DialogType.Error,
-  //         image: Icons.error,
-  //         title: 'Error',
-  //         description:
-  //             'Failed to get the verification status. \nIf this issue persist, please contact support.',
-  //         actions: <Widget>[
-  //           TextButton(
-  //             child: const Text('Close'),
-  //             onPressed: () {
-  //               Navigator.pop(context);
-  //             },
-  //           ),
-  //         ],
-  //       ),
-  //     );
-  //   }
-  //   if (verificationStatus.status == VerificationState.VERIFIED) {
-  //     identityVerified = true;
-  //     setIsIdentityVerified(true);
-  //     Globals().identityVerified.value = true;
-  //     try {
-  //       final data = await getVerificationData(address);
-  //       final firstName = utf8.decode(latin1.encode(data.orgFirstName!));
-  //       final lastName = utf8.decode(latin1.encode(data.orgLastName!));
-  //       final wallets = (await getPkidWallets())
-  //         .where((w) => w.type == WalletType.NATIVE)
-  //         .toList();
-  //       await saveIdentity('$lastName $firstName', data.docIssuingCountry,
-  //           data.docDob, data.docSex, data.idenfyRef, wallets.first.seed );
-  //       Events().emit(IdentityCallbackEvent(type: 'success'));
-  //     } on BadRequest catch (e) {
-  //       setState(() {
-  //         isLoading = false;
-  //       });
-  //       return showDialog(
-  //           context: context,
-  //           builder: (BuildContext context) => CustomDialog(
-  //                 type: DialogType.Warning,
-  //                 image: Icons.warning,
-  //                 title: 'Bad Request',
-  //                 description:
-  //                     '$e \nIf this issue persist, please contact support.',
-  //                 actions: <Widget>[
-  //                   TextButton(
-  //                     child: const Text('Close'),
-  //                     onPressed: () {
-  //                       Navigator.pop(context);
-  //                     },
-  //                   ),
-  //                 ],
-  //               ));
-  //     } on Unauthorized catch (e) {
-  //       setState(() {
-  //         isLoading = false;
-  //       });
-  //       return showDialog(
-  //           context: context,
-  //           builder: (BuildContext context) => CustomDialog(
-  //                 type: DialogType.Warning,
-  //                 image: Icons.warning,
-  //                 title: 'Unauthorized',
-  //                 description:
-  //                     '$e \nIf this issue persist, please contact support.',
-  //                 actions: <Widget>[
-  //                   TextButton(
-  //                     child: const Text('Close'),
-  //                     onPressed: () {
-  //                       Navigator.pop(context);
-  //                     },
-  //                   ),
-  //                 ],
-  //               ));
-  //     } catch (e) {
-  //       setState(() {
-  //         isLoading = false;
-  //       });
-  //       logger.e(e);
-  //       return showDialog(
-  //         context: context,
-  //         builder: (BuildContext context) => CustomDialog(
-  //           type: DialogType.Error,
-  //           image: Icons.error,
-  //           title: 'Error',
-  //           description:
-  //               'Failed to process the verification details. \nIf this issue persist, please contact support.',
-  //           actions: <Widget>[
-  //             TextButton(
-  //               child: const Text('Close'),
-  //               onPressed: () {
-  //                 Navigator.pop(context);
-  //               },
-  //             ),
-  //           ],
-  //         ),
-  //       );
-  //     }
-  //   } else {
-  //     identityVerified = false;
-  //     setIsIdentityVerified(false);
-  //     Globals().identityVerified.value = false;
-  //     Events().emit(IdentityCallbackEvent(type: 'failed'));
-  //   }
-  //   setState(() {});
-  // }
-
-  Widget _pleaseWait() {
-    return Dialog(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          const SizedBox(
-            height: 10,
-          ),
-          CircularProgressIndicator(
-            color: Theme.of(context).colorScheme.primary,
-          ),
-          const SizedBox(
-            height: 10,
-          ),
-          Text(
-            'One moment please',
-            style: Theme.of(context)
-                .textTheme
-                .bodyMedium!
-                .copyWith(color: Theme.of(context).colorScheme.onSurface),
-          ),
-          const SizedBox(
-            height: 10,
-          ),
-        ],
-      ),
-    );
   }
 
   Future _loadingDialog() {
@@ -992,19 +520,6 @@ class _IdentityVerificationScreenState
                             child: Column(
                                 mainAxisAlignment: MainAxisAlignment.center,
                                 children: <Widget>[
-                                  Row(
-                                    children: <Widget>[
-                                      Expanded(
-                                        child: Text(
-                                          step == 3
-                                              ? 'Identity'
-                                              : (text.isEmpty
-                                                  ? 'Unknown'
-                                                  : text),
-                                        ),
-                                      )
-                                    ],
-                                  ),
                                   step == 2 &&
                                           Globals().hidePhoneButton.value ==
                                               true
@@ -1054,20 +569,6 @@ class _IdentityVerificationScreenState
                                         await verifyPhone();
                                       }
                                       break;
-
-                                    // Verify identity
-                                    case 3:
-                                      {
-                                        // await verifyIdentityProcess(
-                                        //     context: context,
-                                        //     setIdentityProcess: (value) =>
-                                        //         setState(() =>
-                                        //             isInIdentityProcess =
-                                        //                 value),
-                                        //     setLoading: (value) => setState(
-                                        //         () => isLoading = value));
-                                      }
-                                      break;
                                     default:
                                       {}
                                       break;
@@ -1100,12 +601,6 @@ class _IdentityVerificationScreenState
           if (step == 1) {
             return _changeEmailDialog(false);
           }
-          // Only make this section clickable if it is Identity Verification + Current Phase
-          if (step != 3) {
-            return;
-          }
-
-          return showIdentityDetails();
         },
         child: Column(children: [
           Padding(
@@ -1167,255 +662,12 @@ class _IdentityVerificationScreenState
                                     ),
                                   ])
                             : const Column(),
-                        step == 3
-                            ? const Column(
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                    Padding(
-                                      padding: EdgeInsets.only(left: 15),
-                                      child: Icon(
-                                        Icons.chevron_right,
-                                      ),
-                                    ),
-                                  ])
-                            : const Column()
                       ],
                     ),
                   ],
                 ),
               ))
         ]));
-  }
-
-  Future<dynamic> showIdentityDetails() {
-    return showDialog(
-        context: context,
-        builder: (BuildContext context) => Dialog(
-              child: FutureBuilder(
-                future: getIdentity(),
-                builder: (BuildContext customContext,
-                    AsyncSnapshot<dynamic> snapshot) {
-                  if (!snapshot.hasData) {
-                    return _pleaseWait();
-                  }
-                  String name = snapshot.data['identityName'];
-                  return Column(
-                    mainAxisSize: MainAxisSize.min,
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const SizedBox(
-                        height: 10,
-                      ),
-                      Container(
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: 15, vertical: 10),
-                          child: Column(
-                            children: [
-                              Text(
-                                'ID CARD',
-                                style: Theme.of(context)
-                                    .textTheme
-                                    .titleLarge!
-                                    .copyWith(
-                                        fontWeight: FontWeight.bold,
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .onSecondaryContainer),
-                                textAlign: TextAlign.center,
-                              ),
-                              const SizedBox(height: 5),
-                              Row(children: [
-                                Text(
-                                  'Your own personal KYC ID CARD',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .titleMedium!
-                                      .copyWith(
-                                          color: Theme.of(context)
-                                              .colorScheme
-                                              .onSecondaryContainer),
-                                ),
-                              ]),
-                            ],
-                          )),
-                      Container(
-                        color: Theme.of(context).colorScheme.secondaryContainer,
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 15, vertical: 20),
-                        child: Column(
-                          children: [
-                            Row(
-                              children: [
-                                Text(
-                                  'Full name',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .primary,
-                                      ),
-                                )
-                              ],
-                            ),
-                            Row(
-                              children: [
-                                Text(
-                                  name,
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .onSecondaryContainer,
-                                      ),
-                                )
-                              ],
-                            )
-                          ],
-                        ),
-                      ),
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 15, vertical: 20),
-                        child: Column(
-                          children: [
-                            Row(
-                              children: [
-                                Text(
-                                  'Birthday',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .primary,
-                                      ),
-                                ),
-                              ],
-                            ),
-                            Row(
-                              children: [
-                                Text(
-                                  snapshot.data['identityDOB'] != 'None'
-                                      ? snapshot.data['identityDOB']
-                                      : 'Unknown',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .onSecondaryContainer,
-                                      ),
-                                )
-                              ],
-                            )
-                          ],
-                        ),
-                      ),
-                      Container(
-                        color: Theme.of(context).colorScheme.secondaryContainer,
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 15, vertical: 20),
-                        child: Column(
-                          children: [
-                            Row(
-                              children: [
-                                Text(
-                                  'Country',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .primary,
-                                      ),
-                                )
-                              ],
-                            ),
-                            Row(
-                              children: [
-                                Text(
-                                  snapshot.data['identityCountry'] != 'None'
-                                      ? snapshot.data['identityCountry']
-                                      : 'Unknown',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .onSecondaryContainer,
-                                      ),
-                                )
-                              ],
-                            )
-                          ],
-                        ),
-                      ),
-                      Container(
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 15, vertical: 20),
-                        child: Column(
-                          children: [
-                            Row(
-                              children: [
-                                Text(
-                                  'Gender',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .primary,
-                                      ),
-                                )
-                              ],
-                            ),
-                            Row(
-                              children: [
-                                Text(
-                                  snapshot.data['identityGender'] != 'None'
-                                      ? snapshot.data['identityGender']
-                                      : 'Unknown',
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .bodyMedium!
-                                      .copyWith(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .onSecondaryContainer,
-                                      ),
-                                )
-                              ],
-                            )
-                          ],
-                        ),
-                      ),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          TextButton(
-                              onPressed: () {
-                                Navigator.pop(customContext);
-                              },
-                              child: const Text('OK')),
-                          const SizedBox(
-                            height: 10,
-                          ),
-                        ],
-                      )
-                    ],
-                  );
-                },
-              ),
-            ));
   }
 
   Future<dynamic> resendEmailDialog(context) {
